Chore: Refactor API to use Flask Blueprints
Migrated the Flask application instance to a Blueprint named api. This change allows the API to be modular and imported into a main app.py file, improving code organization and scalability.
The endpoints /detect and /scale were refactored to be part of the new api blueprint.

Removed the if __name__ == '__main__': block as the application is now intended to be run by a separate file (e.g., app.py).