Refactor: Adjust minor scale chord test
This commit refactors the `test_get_relative_key function in `tests/test_scales_generator.py`.

The previous `scales_generator.py` was using ‘m’ suffix to indicate when a relative scale was minor. Then, it was modified just to show the root of the relative scale because wether it’s major or minor could be inferred from the actual scale.